fix scheduler problem in light_schedule.
Previously, light_scheduler.py used the Python native scheduler library to schedule stimulus events, but whatever reason this became problematic for large values. With this update, and because we don't need the parallel processing potential of the scheduler library, we programmed the scheduling ourselves using time.sleep in between stimulus bouts.